+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 
 ## 0.1.4
 
-- Upgraded dependency to Pointy Castle's 2.0.0 release.
-
+- Upgraded dependency to Pointy Castle's new 2.0.0 release.
 
 ## 0.1.3